+++ index.js
@@ -7,6 +7,7 @@ const ResourcesApi = require('./src/resourcesApi');
 const AccountsApi = require('./src/accountsApi');
 const RolesApi = require('./src/rolesApi');
 const ConnectionsApi = require('./src/connectionsApi');
+const ExtensionsApi = require('./src/extensionsApi');
 const TenantsApi = require('./src/tenantsApi');
 
 class AuthressClient {
@@ -23,6 +24,7 @@ class AuthressClient {
     this.accounts = new AccountsApi(this.httpClient);
     this.roles = new RolesApi(this.httpClient);
     this.connections = new ConnectionsApi(this.httpClient);
+    this.extensions = new ExtensionsApi(this.httpClient);
     this.tenants = new TenantsApi(this.httpClient);
   }